+++ lib/spiral-galaxy.rb
@@ -9,6 +9,11 @@ require 'coffee_script'
 require 'less'
 
 require 'angular-commons-middleware'
+require 'auth-client'
+require 'devcenter-client'
 
 require "spiral-galaxy/version"
+require "spiral-galaxy/connection"
 require "spiral-galaxy/auth_backend"
+require "spiral-galaxy/canvas_redirect"
+require "spiral-galaxy/app"